Extracted token functions from user.ts
