+++ user/src/models/user.ts
@@ -1,4 +1,3 @@
-import jwt from 'jsonwebtoken';
 import { Document, model, Model, Schema, Types } from 'mongoose';
 import { createLocation, getLocationID, locationIsValid } from './location';
 import { hashPassword } from './password';
@@ -51,25 +50,4 @@ async function createUser(user: any): Promise<Document> {
     });
 }
 
-function assignNewToken(user: Document): string {
-    const newToken: string = generateToken(user);
-    user.set('token', newToken);
-    return newToken;
-}
-
-function generateToken(user: Document): string {
-    let token: string = '';
-
-    const secretKey: string | undefined = process.env.JWT_SECRET_KEY;
-    const payload: object = { username: user.get('username'), iat: Date.now() };
-
-    if (secretKey) {
-        token = jwt.sign(payload, secretKey, { expiresIn: '168h' });
-    } else {
-        throw new Error('Secret key is not defined.');
-    }
-
-    return token;
-}
-
-export { createUser, assignNewToken };
+export { createUser };
